Move filing status field and remove step
